mini3/requirements.txt
streamlit==1.38.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
pydicom>=2.4.0

